backend/tests/requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-mock==3.12.0
faker==20.1.0
responses==0.24.1
freezegun==1.4.0
httpx>=0.27.0
pillow==10.1.0
